Fixed non-english characters coming as _ and streamlined logic in download service (Fixes #4589)
