+++ UI/Web/src/app/shared/_models/download.ts
@@ -1,4 +1,4 @@
-import {HttpEvent, HttpEventType, HttpHeaders, HttpProgressEvent, HttpResponse} from "@angular/common/http";
+import {HttpEvent, HttpEventType, HttpProgressEvent, HttpResponse} from "@angular/common/http";
 import {Observable} from "rxjs";
 import {scan} from "rxjs/operators";
 
@@ -44,13 +44,13 @@ export function download(saver?: (b: Blob, filename: string) => void): (source: 
             }
             if (isHttpResponse(event)) {
               if (saver && event.body) {
-                saver(event.body, getFilename(event.headers, ''))
+                saver(event.body, parseContentDisposition(event.headers.get('content-disposition') || '', ''))
               }
               return {
                 progress: 100,
                 state: 'DONE',
                 content: event.body,
-                filename: getFilename(event.headers, ''),
+                filename: parseContentDisposition(event.headers.get('content-disposition') || '', ''),
               }
             }
             return previous;
@@ -61,16 +61,44 @@ export function download(saver?: (b: Blob, filename: string) => void): (source: 
   }
 
 
-function getFilename(headers: HttpHeaders, defaultName: string) {
-    const tokens = (headers.get('content-disposition') || '').split(';');
+/**
+ * Parse Content-Disposition header to extract filename, with fallback.
+ * Prefers filename*=UTF-8'' (RFC 5987) for non-ASCII filenames.
+ */
+export function parseContentDisposition(header: string, fallbackName: string): string {
+    if (!header) return fallbackName || 'download';
+    const tokens = header.split(';');
+
+    if (tokens.length < 2) return fallbackName || 'download';
+
+    // Prefer filename*=UTF-8'' (RFC 5987) for non-ASCII filenames
+    const starToken = tokens.find(t => t.trim().toLowerCase().startsWith('filename*='));
+    if (starToken) {
+      const parts = starToken.trim().split("''");
+      if (parts.length === 2 && parts[1]) {
+        try {
+          const filename = decodeURIComponent(parts[1]);
+          if (filename.startsWith('download_') || filename.startsWith('kavita_download_')) {
+            const ext = filename.substring(filename.lastIndexOf('.'), filename.length);
+            if (fallbackName) return fallbackName + ext;
+            return filename.replace('kavita_', '').replace('download_', '');
+          }
+          return filename;
+        } catch { /* fall through to filename= */ }
+      }
+    }
+
     let filename = tokens[1].replace('filename=', '').replace(/"/ig, '').trim();
+
     if (filename.startsWith('download_') || filename.startsWith('kavita_download_')) {
       const ext = filename.substring(filename.lastIndexOf('.'), filename.length);
-      if (defaultName !== '') {
-        return defaultName + ext;
-      }
+      if (fallbackName) return fallbackName + ext;
       return filename.replace('kavita_', '').replace('download_', '');
-      //return defaultName + ext;
     }
-    return filename;
+
+    try {
+      return decodeURIComponent(filename) || fallbackName || 'download';
+    } catch {
+      return filename || fallbackName || 'download';
+    }
   }
